requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
google-generativeai>=0.3.0
PyYAML>=6.0 